requirements.txt
cassandra-driver==3.25.0
click==8.1.3
geomet==0.2.1.post1
importlib-metadata==4.12.0
mysql-connector-python==8.0.30
protobuf==3.20.1
six==1.16.0
typing_extensions==4.3.0
zipp==3.8.1
